fix deflate/deflate.c

+++ deflate/no_compression.h
@@ -13,7 +13,7 @@
  * @param last_block 	is this last block ?
  * @param bs_out 	output bit stream
  */
-void deflate_no_compression_compress(uint8_t *block, uint32_t len, int last_block, struct bit_stream *bs_out);
+void deflate_no_compression_compress(uint8_t *block, uint16_t len, int last_block, struct bit_stream *bs_out);
 
 /**
  * @brief Uncompress a block.
